+++ src/SQLProvider/Providers.MsSqlServer.Dynamic.fs
@@ -4,19 +4,62 @@ open System
 open System.Collections.Concurrent
 open System.Collections.Generic
 open System.Data
-open System.Data.SqlClient
 open FSharp.Data.Sql
 open FSharp.Data.Sql.Transactions
 open FSharp.Data.Sql.Schema
 open FSharp.Data.Sql.Common
 
-module MSSqlServer =
+module MSSqlServerDynamic =
+    let mutable resolutionPath = String.Empty
+    let mutable referencedAssemblies = [||]
+
+    let assemblyNames = [
+        "Microsoft.Data.SqlClient.dll"; "System.Data.SqlClient.dll"
+    ]
+
+    let assembly =
+        lazy Reflection.tryLoadAssemblyFrom resolutionPath referencedAssemblies assemblyNames
+
+    let findType name =
+        match assembly.Value with
+        | Choice1Of2(assembly) -> 
+            let types = 
+                try assembly.GetTypes() 
+                with | :? System.Reflection.ReflectionTypeLoadException as e ->
+                    let msgs = e.LoaderExceptions |> Seq.map(fun e -> e.GetBaseException().Message) |> Seq.distinct
+                    let details = "Details: " + Environment.NewLine + String.Join(Environment.NewLine, msgs)
+                    failwith (e.Message + Environment.NewLine + details)
+            match types |> Array.tryFind(fun t -> t.Name = name) with
+            | Some t -> t
+            | None -> failwith ("Assembly " + assembly.FullName + " found, but it didn't contain expected type " + name +
+                                 Environment.NewLine + "Tired to load a dll: " + assembly.CodeBase)
+
+        | Choice2Of2(paths, errors) ->
+           let details = 
+                match errors with 
+                | [] -> "" 
+                | x -> Environment.NewLine + "Details: " + Environment.NewLine + String.Join(Environment.NewLine, x)
+           failwithf "Unable to resolve assemblies. One of %s (e.g. from Nuget package Microsoft.Data.SqlClient) must exist in the paths: %s %s %s"
+                (String.Join(", ", assemblyNames |> List.toArray))
+                Environment.NewLine
+                (String.Join(Environment.NewLine, paths |> Seq.filter(fun p -> not(String.IsNullOrEmpty p))))
+                details
+
+    let connectionType =  lazy (findType "SqlConnection")
+    let commandType =     lazy (findType "SqlCommand")
+    let parameterType =   lazy (findType "SqlParameter")
+    let enumType =        lazy (findType "SqlDbType")
+    let getSchemaMethod = lazy (connectionType.Value.GetMethod("GetSchema",[|typeof<string>; typeof<string[]>|]))
+
     let getSchema name (args:string[]) (con:IDbConnection) =
-        let con = (con :?> SqlConnection)
         if con.State <> ConnectionState.Open then con.Open()
-        let res = con.GetSchema(name, args)
+        let res = getSchemaMethod.Value.Invoke(con,[|name; args|]) :?> DataTable
         con.Close()
         res
+
+    let parseDbType dbTypeName =
+        try Some(Enum.Parse(enumType.Value, dbTypeName) |> unbox<int>)
+        with _ -> None
 
     let mutable typeMappings = []
     let mutable findClrType : (string -> TypeMapping option)  = fun _ -> failwith "!"
@@ -26,13 +69,19 @@ module MSSqlServer =
         let dt = getSchema "DataTypes" [||] con
 
         let getDbType(providerType:int) =
-            let p = new SqlParameter()
-            if providerType = 31
-            then p.SqlDbType <- SqlDbType.DateTime
-            else if providerType = 32
-            then p.SqlDbType <- SqlDbType.Time
-            else p.SqlDbType <- (Enum.ToObject(typeof<SqlDbType>, providerType) :?> SqlDbType)
-            p.DbType
+            let pv = parameterType.Value
+            let p = Activator.CreateInstance(pv,[||]) :?> IDbDataParameter
+            let setter = pv.GetProperty("SqlDbType").GetSetMethod()
+            let dbTypeGetter = pv.GetProperty("DbType").GetGetMethod()
+            setter.Invoke(p, [| 
+                (if providerType = 31
+                 then (match parseDbType "DateTime" with Some x -> x | None -> providerType)
+                 else if providerType = 32
+                 then (match parseDbType "Time" with Some x -> x | None -> providerType)
+                 else providerType)
+            |]) |> ignore
+
+            dbTypeGetter.Invoke(p, [||]) :?> DbType
 
         let getClrType (input:string) =
             let t = Type.GetType input
@@ -75,9 +124,24 @@ module MSSqlServer =
         findClrType <- clrMappings.TryFind
         findDbType <- dbMappings.TryFind
 
-    let createConnection connectionString = new SqlConnection(connectionString) :> IDbConnection
+    let createConnection connectionString =
+        try
+            Activator.CreateInstance(connectionType.Value,[|box connectionString|]) :?> IDbConnection
+        with
+        | :? System.Reflection.ReflectionTypeLoadException as ex ->
+            let errorfiles = ex.LoaderExceptions |> Array.map(fun e -> e.GetBaseException().Message) |> Seq.distinct |> Seq.toArray
+            let msg = ex.GetBaseException().Message + "\r\n" + String.Join("\r\n", errorfiles)
+            raise(new System.Reflection.TargetInvocationException(msg, ex))
+        | :? System.Reflection.TargetInvocationException as ex when (ex.InnerException <> null && ex.InnerException :? DllNotFoundException) ->
+            let msg = ex.GetBaseException().Message + ", Path: " + (System.IO.Path.GetFullPath resolutionPath)
+            raise(new System.Reflection.TargetInvocationException(msg, ex))
+        | :? System.TypeInitializationException as te when (te.InnerException :? System.Reflection.TargetInvocationException) ->
+            let ex = te.InnerException :?> System.Reflection.TargetInvocationException
+            let msg = ex.GetBaseException().Message + ", Path: " + (System.IO.Path.GetFullPath resolutionPath)
+            raise(new System.Reflection.TargetInvocationException(msg, ex.InnerException)) 
 
-    let createCommand commandText (connection:IDbConnection) = new SqlCommand(commandText, downcast connection) :> IDbCommand
+    let createCommand commandText (connection:IDbConnection) =
+        Activator.CreateInstance(commandType.Value,[|box commandText;box connection|]) :?> IDbCommand
 
     let dbUnbox<'a> (v:obj) : 'a =
         if Convert.IsDBNull(v) then Unchecked.defaultof<'a> else unbox v
@@ -91,13 +155,12 @@ module MSSqlServer =
         con.Close(); result
 
     let executeSql sql (con:IDbConnection) =
-        use com = new SqlCommand(sql,con:?>SqlConnection)
+        use com = Activator.CreateInstance(commandType.Value,[|box sql;box con|]) :?> IDbCommand
         com.ExecuteReader()
 
     let readParameter (parameter:IDbDataParameter) =
         if parameter <> null then
-            let par = parameter :?> SqlParameter
-            par.Value
+            parameter.Value
         else null
         
     let readInOutParameterFromCommand name (com:IDbCommand) = 
@@ -106,28 +169,39 @@ module MSSqlServer =
         match com.Parameters.Item name :?> IDataParameter with
         | p when p.Direction = ParameterDirection.InputOutput -> p.ParameterName, p.Value
         | p -> failwithf "Unsupported direction %A for parameter %A" p.Direction p.ParameterName
-            
+
     let createOpenParameter(name,v:obj)= 
-        let p = SqlParameter(name,v)
+        let parameterType = parameterType.Value
+        let udtTypeSetter =
+            parameterType.GetProperty("UdtTypeName").GetSetMethod()
+
+        let p = Activator.CreateInstance(parameterType,[|box name;v|]) :?> IDbDataParameter
         match v.GetType().FullName with
-        | "Microsoft.SqlServer.Types.SqlGeometry" -> p.UdtTypeName <- "Geometry"
-        | "Microsoft.SqlServer.Types.SqlGeography" -> p.UdtTypeName <- "Geography"
-        | "Microsoft.SqlServer.Types.SqlHierarchyId" -> p.UdtTypeName <- "HierarchyId"
+        | "Microsoft.SqlServer.Types.SqlGeometry" -> udtTypeSetter.Invoke(p, [| "Geometry" |]) |> ignore
+        | "Microsoft.SqlServer.Types.SqlGeography" -> udtTypeSetter.Invoke(p, [| "Geography" |]) |> ignore
+        | "Microsoft.SqlServer.Types.SqlHierarchyId" -> udtTypeSetter.Invoke(p, [| "HierarchyId" |]) |> ignore
         | _ -> ()
         p
 
     let createCommandParameter (param:QueryParameter) (value:obj) =
-        let p = SqlParameter(param.Name,value)
+        let parameterType = parameterType.Value
+        let sqlDbTypeSetter =
+            parameterType.GetProperty("SqlDbType").GetSetMethod()
+        let udtTypeSetter =
+            parameterType.GetProperty("UdtTypeName").GetSetMethod()
+
+        let p = Activator.CreateInstance(parameterType,[|box param.Name;value|]) :?> IDbDataParameter
         p.DbType <- param.TypeMapping.DbType
-        Option.iter (fun (t:int) -> p.SqlDbType <- Enum.ToObject(typeof<SqlDbType>, t) :?> SqlDbType) param.TypeMapping.ProviderType
+        Option.iter (fun (t:int) ->
+            sqlDbTypeSetter.Invoke(p, [| t |]) |> ignore) param.TypeMapping.ProviderType
         p.Direction <- param.Direction
         Option.iter (fun l -> p.Size <- l) param.Length
         match param.TypeMapping.ProviderTypeName with
-        | Some "Microsoft.SqlServer.Types.SqlGeometry" -> p.UdtTypeName <- "Geometry"
-        | Some "Microsoft.SqlServer.Types.SqlGeography" -> p.UdtTypeName <- "Geography"
-        | Some "Microsoft.SqlServer.Types.SqlHierarchyId" -> p.UdtTypeName <- "HierarchyId"
+        | Some "Microsoft.SqlServer.Types.SqlGeometry" -> udtTypeSetter.Invoke(p, [| "Geometry" |]) |> ignore
+        | Some "Microsoft.SqlServer.Types.SqlGeography" -> udtTypeSetter.Invoke(p, [| "Geography" |]) |> ignore
+        | Some "Microsoft.SqlServer.Types.SqlHierarchyId" -> udtTypeSetter.Invoke(p, [| "HierarchyId" |]) |> ignore
         | _ -> ()
-        p :> IDbDataParameter
+        p
 
     let getSprocReturnCols (con: IDbConnection) (sname: SprocName) (sparams: QueryParameter list) =
         let parameterStr =
@@ -204,7 +278,6 @@ module MSSqlServer =
         |> Seq.toList
 
     let getSprocs (con: IDbConnection) =
-        let con = (con :?> SqlConnection)
 
         let tableValued =
             Sql.executeSqlAsDataTable
@@ -296,7 +369,7 @@ module MSSqlServer =
         | [|retCol|] ->
             match retCol.TypeMapping.ProviderTypeName with
             | Some "cursor" ->
-                use reader = com.ExecuteReader() :?> SqlDataReader
+                use reader = com.ExecuteReader()
                 let result = SingleResultSet(retCol.Name, Sql.dataReaderToArray reader)
                 reader.NextResult() |> ignore
                 result
@@ -306,7 +379,7 @@ module MSSqlServer =
                 | Some(_,name,p) -> Scalar(name, readParameter p)
                 | None -> readInOutParameterFromCommand retCol.Name com |> Scalar
         | cols ->
-            use reader = com.ExecuteReader() :?> SqlDataReader
+            use reader = com.ExecuteReader()
             Set(cols |> Array.map (processReturnColumn com reader))
 
 
@@ -322,7 +395,7 @@ module MSSqlServer =
                 match retCol.TypeMapping.ProviderTypeName with
                 | Some "cursor" ->
                     use! readera = com.ExecuteReaderAsync() |> Async.AwaitTask
-                    let reader = readera :?> SqlDataReader
+                    let reader = readera
                     let! r = Sql.dataReaderToArrayAsync reader
                     let result = SingleResultSet(retCol.Name, r)
                     let! _ = reader.NextResultAsync() |> Async.AwaitTask
@@ -337,14 +410,9 @@ module MSSqlServer =
                 let! r = cols |> Array.toList |> Sql.evaluateOneByOne (processReturnColumnAsync com reader)
                 return Set(r |> List.toArray)
         }
-        
-type internal MSSQLPagingCompatibility =
-  // SQL SERVER versions since 2012
-  | Offset = 0
-  // SQL SERVER versions prior to 2012
-  | RowNumber = 1
 
-type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
+open MSSqlServerDynamic
+type internal MSSqlServerDynamicProvider(resolutionPath, contextSchemaPath, referencedAssemblies, tableNames:string) as this =
     let schemaCache = SchemaCache.LoadOrEmpty(contextSchemaPath)
     let myLock = new Object()
     
@@ -361,15 +429,15 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
     let createInsertCommand (con:IDbConnection) (sb:Text.StringBuilder) (entity:SqlEntity) =
         let (~~) (t:string) = sb.Append t |> ignore
 
-        let cmd = new SqlCommand()
-        cmd.Connection <- con :?> SqlConnection
+        let cmd = (this :> ISqlProvider).CreateCommand(con,"")
+        cmd.Connection <- con
         let haspk = schemaCache.PrimaryKeys.ContainsKey(entity.Table.FullName)
         let pk = if haspk then schemaCache.PrimaryKeys.[entity.Table.FullName] else []
         let columnNames, values =
             (([],0),entity.ColumnValues)
             ||> Seq.fold(fun (out,i) (k,v) ->
                 let name = sprintf "@param%i" i
-                let p = MSSqlServer.createOpenParameter(name,v)
+                let p = createOpenParameter(name,v)
                 (sprintf "[%s]" k,p)::out,i+1)
             |> fst
             |> List.rev
@@ -396,17 +464,18 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                     (String.Join(",",columnNames))
                     (String.Join(",",values |> Array.map(fun p -> p.ParameterName))))
 
-        cmd.Parameters.AddRange(values)
+        values |> Array.iter(fun v -> cmd.Parameters.Add v |> ignore)
         cmd.CommandText <- sb.ToString()
         cmd
 
     let createUpdateCommand (con:IDbConnection) (sb:Text.StringBuilder) (entity:SqlEntity) (changedColumns:string list) =
         let (~~) (t:string) = sb.Append t |> ignore
 
-        let cmd = new SqlCommand()
-        cmd.Connection <- con :?> SqlConnection
+        let cmd = (this :> ISqlProvider).CreateCommand(con,"")
+        cmd.Connection <- con
         let haspk = schemaCache.PrimaryKeys.ContainsKey(entity.Table.FullName)
         let pk = if haspk then schemaCache.PrimaryKeys.[entity.Table.FullName] else []
+
         sb.Clear() |> ignore
         match pk with
         | [x] when changedColumns |> List.exists ((=)x)
@@ -425,9 +494,9 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                 let p =
                     match entity.GetColumnOption<obj> col with
                     | Some v ->
-                        let p = MSSqlServer.createOpenParameter(name,v)
+                        let p = createOpenParameter(name,v)
                         p
-                    | None -> MSSqlServer.createOpenParameter(name,DBNull.Value)
+                    | None -> createOpenParameter(name, DBNull.Value)
                 (col,p)::out,i+1)
             |> fst
             |> List.rev
@@ -441,9 +510,9 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                 (String.Join(",", data |> Array.map(fun (c,p) -> sprintf "[%s] = %s" c p.ParameterName ) )))
             ~~(String.Join(" AND ", ks |> List.mapi(fun i k -> (sprintf "[%s] = @pk%i" k i))))
 
-        cmd.Parameters.AddRange(data |> Array.map snd)
+        data |> Array.map snd |> Array.iter(fun v ->cmd.Parameters.Add v |> ignore)
         pkValues |> List.iteri(fun i pkValue ->
-            let pkParam = MSSqlServer.createOpenParameter(("@pk"+i.ToString()), pkValue)
+            let pkParam = createOpenParameter("@pk"+i.ToString(),pkValue)
             cmd.Parameters.Add pkParam |> ignore)
         cmd.CommandText <- sb.ToString()
         cmd
@@ -451,8 +520,8 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
     let createDeleteCommand (con:IDbConnection) (sb:Text.StringBuilder) (entity:SqlEntity) =
         let (~~) (t:string) = sb.Append t |> ignore
 
-        let cmd = new SqlCommand()
-        cmd.Connection <- con :?> SqlConnection
+        let cmd = (this :> ISqlProvider).CreateCommand(con,"")
+        cmd.Connection <- con 
         sb.Clear() |> ignore
         let haspk = schemaCache.PrimaryKeys.ContainsKey(entity.Table.FullName)
         let pk = if haspk then schemaCache.PrimaryKeys.[entity.Table.FullName] else []
@@ -463,7 +532,8 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
             | v -> v
 
         pkValues |> List.iteri(fun i pkValue ->
-            cmd.Parameters.AddWithValue(("@id"+i.ToString()),pkValue) |> ignore)
+            let p = (this :> ISqlProvider).CreateCommandParameter(QueryParameter.Create("@id"+i.ToString(), i),pkValue)
+            cmd.Parameters.Add(p) |> ignore)
 
         match pk with
         | [] -> ()
@@ -473,6 +543,10 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
 
         cmd.CommandText <- sb.ToString()
         cmd
+
+    do
+        MSSqlServerDynamic.resolutionPath <- resolutionPath
+        MSSqlServerDynamic.referencedAssemblies <- referencedAssemblies
 
     interface ISqlProvider with
         member __.GetLockObject() = myLock
@@ -485,8 +559,9 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                                                         and (sep.minor_id is null or sep.minor_id =0)
                                                         and sep.name = 'MS_Description'
                 where st.name = @TableName"""
-            use com = new SqlCommand(baseq,con:?>SqlConnection)
-            com.Parameters.AddWithValue("@TableName",tn) |> ignore
+            use com = (this:>ISqlProvider).CreateCommand(con,baseq)
+            com.Parameters.Add((this:>ISqlProvider).CreateCommandParameter(QueryParameter.Create("@TableName", 0), tn)) |> ignore
+
             if con.State <> ConnectionState.Open then con.Open()
             use reader = com.ExecuteReader()
             if reader.Read() then
@@ -507,9 +582,9 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                                                         and sep.name = 'MS_Description'
                 where st.name = @TableName
                 and sc.name = @ColumnName"""
-            use com = new SqlCommand(baseq,con:?>SqlConnection)
-            com.Parameters.AddWithValue("@TableName",tn) |> ignore
-            com.Parameters.AddWithValue("@ColumnName",columnName) |> ignore
+            use com = (this:>ISqlProvider).CreateCommand(con,baseq)
+            com.Parameters.Add((this:>ISqlProvider).CreateCommandParameter(QueryParameter.Create("@TableName", 0), tn)) |> ignore
+            com.Parameters.Add((this:>ISqlProvider).CreateCommandParameter(QueryParameter.Create("@ColumnName", 1), columnName)) |> ignore
             if con.State <> ConnectionState.Open then con.Open()
             use reader = com.ExecuteReader()
             if reader.Read() then
@@ -518,12 +593,12 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                 with 
                 | :? InvalidCastException -> ""
             else ""
-        member __.CreateConnection(connectionString) = MSSqlServer.createConnection connectionString
-        member __.CreateCommand(connection,commandText) = MSSqlServer.createCommand commandText connection
-        member __.CreateCommandParameter(param, value) = MSSqlServer.createCommandParameter param value
-        member __.ExecuteSprocCommand(con, inputParameters, returnCols, values:obj array) = MSSqlServer.executeSprocCommand con inputParameters returnCols values
-        member __.ExecuteSprocCommandAsync(con, inputParameters, returnCols, values:obj array) = MSSqlServer.executeSprocCommandAsync con inputParameters returnCols values
-        member __.CreateTypeMappings(con) = MSSqlServer.createTypeMappings con
+        member __.CreateConnection(connectionString) = MSSqlServerDynamic.createConnection connectionString
+        member __.CreateCommand(connection,commandText) = MSSqlServerDynamic.createCommand commandText connection
+        member __.CreateCommandParameter(param, value) = MSSqlServerDynamic.createCommandParameter param value
+        member __.ExecuteSprocCommand(con, inputParameters, returnCols, values:obj array) = MSSqlServerDynamic.executeSprocCommand con inputParameters returnCols values
+        member __.ExecuteSprocCommandAsync(con, inputParameters, returnCols, values:obj array) = MSSqlServerDynamic.executeSprocCommandAsync con inputParameters returnCols values
+        member __.CreateTypeMappings(con) = MSSqlServerDynamic.createTypeMappings con
         member __.GetSchemaCache() = schemaCache
         
         member __.GetTables(con,_) =
@@ -531,10 +606,10 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                 match tableNames with 
                 | "" -> ""
                 | x -> " where 1=1 " + (SchemaProjections.buildTableNameWhereFilter "TABLE_NAME" tableNames)
-            MSSqlServer.connect con (fun con ->
-            use reader = MSSqlServer.executeSql ("select TABLE_SCHEMA, TABLE_NAME, TABLE_TYPE from INFORMATION_SCHEMA.TABLES" + tableNamesFilter) con
+            MSSqlServerDynamic.connect con (fun con ->
+            use reader = MSSqlServerDynamic.executeSql ("select TABLE_SCHEMA, TABLE_NAME, TABLE_TYPE from INFORMATION_SCHEMA.TABLES" + tableNamesFilter) con
             [ while reader.Read() do
-                let table ={ Schema = reader.GetSqlString(0).Value ; Name = reader.GetSqlString(1).Value ; Type=reader.GetSqlString(2).Value.ToLower() }
+                let table ={ Schema = reader.GetString(0) ; Name = reader.GetString(1) ; Type=reader.GetString(2).ToLower() }
                 yield schemaCache.Tables.GetOrAdd(table.FullName,table)
                 ])
 
@@ -550,7 +625,10 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
             mssqlVersionCache.GetOrAdd(con.ConnectionString, fun conn ->
                 lazy
                     if con.State <> ConnectionState.Open then con.Open()
-                    let success, version = (con :?> SqlConnection).ServerVersion |> Version.TryParse
+                    let ver = con.GetType().GetProperty("ServerVersion").GetValue(con,null)
+                    if ver = null then Version("12.0")
+                    else
+                    let success, version = ver.ToString() |> Version.TryParse
                     if success then version else Version("12.0")
             ).Value |> ignore 
                 
@@ -581,25 +659,26 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                                              AND c.COLUMN_NAME = pk.COLUMN_NAME
                                  WHERE c.TABLE_SCHEMA = @schema AND c.TABLE_NAME = @table
                                  ORDER BY c.TABLE_SCHEMA,c.TABLE_NAME, c.ORDINAL_POSITION"
-               use com = new SqlCommand(baseQuery,con:?>SqlConnection)
-               com.Parameters.AddWithValue("@schema",table.Schema) |> ignore
-               com.Parameters.AddWithValue("@table",table.Name) |> ignore
+               use com = (this:>ISqlProvider).CreateCommand(con,baseQuery)
+               com.Parameters.Add((this:>ISqlProvider).CreateCommandParameter(QueryParameter.Create("@schema", 0), table.Schema)) |> ignore
+               com.Parameters.Add((this:>ISqlProvider).CreateCommandParameter(QueryParameter.Create("@table", 1), table.Name)) |> ignore
+
                if con.State <> ConnectionState.Open then con.Open()
 
                use reader = com.ExecuteReader()
                let columns =
                    [ while reader.Read() do
-                       let dt = reader.GetSqlString(1).Value
-                       let maxlen = 
-                            let x = reader.GetSqlInt32(2)
-                            if x.IsNull then 0 else (x.Value)
-                       match MSSqlServer.findDbType dt with
+                       let dt = reader.GetString(1)
+                       let maxlen =
+                            if reader.IsDBNull(2) then 0
+                            else reader.GetInt32(2)
+                       match MSSqlServerDynamic.findDbType dt with
                        | Some(m) ->
                            let col =
-                             { Column.Name = reader.GetSqlString(0).Value;
+                             { Column.Name = reader.GetString(0);
                                TypeMapping = m
                                IsNullable = let b = reader.GetString(4) in if b = "YES" then true else false
-                               IsPrimaryKey = if reader.GetSqlString(5).Value = "PRIMARY KEY" then true else false
+                               IsPrimaryKey = if reader.GetString(5) = "PRIMARY KEY" then true else false
                                IsAutonumber = reader.GetInt32(6) = 1
                                HasDefault = reader.GetInt32(7) = 1
                                TypeInfo = if maxlen > 0 then Some (dt + "(" + maxlen.ToString() + ")") else Some dt }
@@ -645,30 +724,31 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                                 AND KCU2.CONSTRAINT_NAME = RC.UNIQUE_CONSTRAINT_NAME
                                 AND KCU2.ORDINAL_POSITION = KCU1.ORDINAL_POSITION "
 
-            let res = MSSqlServer.connect con (fun con ->
+            let res = MSSqlServerDynamic.connect con (fun con ->
                 let baseq1 = sprintf "%s WHERE KCU2.TABLE_NAME = @tblName" baseQuery 
-                use com1 = new SqlCommand(baseq1,con:?>SqlConnection)
-                com1.Parameters.AddWithValue("@tblName",table.Name) |> ignore
+                use com1 = (this:>ISqlProvider).CreateCommand(con,baseq1)
+                com1.Parameters.Add((this:>ISqlProvider).CreateCommandParameter(QueryParameter.Create("@tblName", 0), table.Name)) |> ignore
                 if con.State <> ConnectionState.Open then con.Open()
                 use reader = com1.ExecuteReader()
                 let children =
                     [ while reader.Read() do
-                        yield { Name = reader.GetSqlString(0).Value; PrimaryTable=Table.CreateFullName(reader.GetSqlString(9).Value, reader.GetSqlString(5).Value); PrimaryKey=reader.GetSqlString(6).Value
-                                ForeignTable= Table.CreateFullName(reader.GetSqlString(8).Value, reader.GetSqlString(1).Value); ForeignKey=reader.GetSqlString(2).Value } ]
+                        yield { Name = reader.GetString(0); PrimaryTable=Table.CreateFullName(reader.GetString(9), reader.GetString(5)); PrimaryKey=reader.GetString(6)
+                                ForeignTable= Table.CreateFullName(reader.GetString(8), reader.GetString(1)); ForeignKey=reader.GetString(2) } ]
                 reader.Dispose()
                 let baseq2 = sprintf "%s WHERE KCU1.TABLE_NAME = @tblName" baseQuery
-                use com2 = new SqlCommand(baseq2,con:?>SqlConnection)
-                com2.Parameters.AddWithValue("@tblName",table.Name) |> ignore
+                use com2 = (this:>ISqlProvider).CreateCommand(con,baseq2)
+                com2.Parameters.Add((this:>ISqlProvider).CreateCommandParameter(QueryParameter.Create("@tblName", 0), table.Name)) |> ignore
+
                 if con.State <> ConnectionState.Open then con.Open()
                 use reader = com2.ExecuteReader()
                 let parents =
                     [ while reader.Read() do
-                        yield { Name = reader.GetSqlString(0).Value; PrimaryTable=Table.CreateFullName(reader.GetSqlString(9).Value, reader.GetSqlString(5).Value); PrimaryKey=reader.GetSqlString(6).Value
-                                ForeignTable=Table.CreateFullName(reader.GetSqlString(8).Value, reader.GetSqlString(1).Value); ForeignKey=reader.GetSqlString(2).Value } ]
+                        yield { Name = reader.GetString(0); PrimaryTable=Table.CreateFullName(reader.GetString(9), reader.GetString(5)); PrimaryKey=reader.GetString(6)
+                                ForeignTable=Table.CreateFullName(reader.GetString(8), reader.GetString(1)); ForeignKey=reader.GetString(2) } ]
                 (children,parents))
             res)
 
-        member __.GetSprocs(con) = MSSqlServer.connect con MSSqlServer.getSprocs
+        member __.GetSprocs(con) = MSSqlServerDynamic.connect con MSSqlServerDynamic.getSprocs
         member __.GetIndividualsQueryText(table,amount) = sprintf "SELECT TOP %i * FROM %s" amount table.FullName
         member __.GetIndividualQueryText(table,column) = sprintf "SELECT * FROM [%s].[%s] WHERE [%s].[%s].[%s] = @id" table.Schema table.Name table.Schema table.Name column
 
@@ -682,14 +762,14 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
 
             let createParam (value:obj) =
                 let paramName = nextParam()
-                let p = MSSqlServer.createOpenParameter(paramName,value)
-                p :> IDbDataParameter
+                createOpenParameter(paramName,value)
 
             let fieldParam (value:obj) =
                 let paramName = nextParam()
-                parameters.Add(MSSqlServer.createOpenParameter(paramName,value):> IDbDataParameter)
+                let p = createOpenParameter(paramName,value)
+                parameters.Add(p)
                 paramName
-                            
+
             let mssqlPaging =               
               match mssqlVersionCache.TryGetValue(con.ConnectionString) with
               // SQL 2008 and earlier do not support OFFSET
@@ -1075,7 +1155,7 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                 |> Seq.iter(fun e ->
                     match e._State with
                     | Created ->
-                        let cmd = createInsertCommand con sb e
+                        let cmd = createInsertCommand con sb e :?> System.Data.Common.DbCommand
                         Common.QueryEvents.PublishSqlQueryCol con.ConnectionString cmd.CommandText cmd.Parameters
                         if timeout.IsSome then
                             cmd.CommandTimeout <- timeout.Value
@@ -1083,14 +1163,14 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                         CommonTasks.checkKey schemaCache.PrimaryKeys id e
                         e._State <- Unchanged
                     | Modified fields ->
-                        let cmd = createUpdateCommand con sb e fields
+                        let cmd = createUpdateCommand con sb e fields :?> System.Data.Common.DbCommand
                         Common.QueryEvents.PublishSqlQueryCol con.ConnectionString cmd.CommandText cmd.Parameters
                         if timeout.IsSome then
                             cmd.CommandTimeout <- timeout.Value
                         cmd.ExecuteNonQuery() |> ignore
                         e._State <- Unchanged
                     | Delete ->
-                        let cmd = createDeleteCommand con sb e
+                        let cmd = createDeleteCommand con sb e :?> System.Data.Common.DbCommand
                         Common.QueryEvents.PublishSqlQueryCol con.ConnectionString cmd.CommandText cmd.Parameters
                         if timeout.IsSome then
                             cmd.CommandTimeout <- timeout.Value
@@ -1125,7 +1205,7 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                         match e._State with
                         | Created ->
                             async {
-                                let cmd = createInsertCommand con sb e
+                                let cmd = createInsertCommand con sb e :?> System.Data.Common.DbCommand
                                 Common.QueryEvents.PublishSqlQueryCol con.ConnectionString cmd.CommandText cmd.Parameters
                                 if timeout.IsSome then
                                     cmd.CommandTimeout <- timeout.Value
@@ -1135,7 +1215,7 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                             }
                         | Modified fields ->
                             async {
-                                let cmd = createUpdateCommand con sb e fields
+                                let cmd = createUpdateCommand con sb e fields :?> System.Data.Common.DbCommand
                                 Common.QueryEvents.PublishSqlQueryCol con.ConnectionString cmd.CommandText cmd.Parameters
                                 if timeout.IsSome then
                                     cmd.CommandTimeout <- timeout.Value
@@ -1144,7 +1224,7 @@ type internal MSSqlServerProvider(contextSchemaPath, tableNames:string) =
                             }
                         | Delete ->
                             async {
-                                let cmd = createDeleteCommand con sb e
+                                let cmd = createDeleteCommand con sb e :?> System.Data.Common.DbCommand
                                 Common.QueryEvents.PublishSqlQueryCol con.ConnectionString cmd.CommandText cmd.Parameters
                                 if timeout.IsSome then
                                     cmd.CommandTimeout <- timeout.Value