+++ src/SQLProvider/SqlRuntime.DataContext.fs
@@ -14,8 +14,10 @@ module internal ProviderBuilder =
     open FSharp.Data.Sql.Providers
 
     let createProvider vendor resolutionPath referencedAssemblies runtimeAssembly owner tableNames contextSchemaPath odbcquote sqliteLibrary =
+        let referencedAssemblies = Array.append [|runtimeAssembly|] referencedAssemblies
         match vendor with
         | DatabaseProviderTypes.MSSQLSERVER -> MSSqlServerProvider(contextSchemaPath, tableNames) :> ISqlProvider
+        | DatabaseProviderTypes.MSSQLSERVER_DYNAMIC -> MSSqlServerDynamicProvider(resolutionPath, contextSchemaPath, referencedAssemblies, tableNames) :> ISqlProvider
         | DatabaseProviderTypes.SQLITE -> SQLiteProvider(resolutionPath, contextSchemaPath, referencedAssemblies, runtimeAssembly, sqliteLibrary) :> ISqlProvider
         | DatabaseProviderTypes.POSTGRESQL -> PostgresqlProvider(resolutionPath, contextSchemaPath, owner, referencedAssemblies) :> ISqlProvider
         | DatabaseProviderTypes.MYSQL -> MySqlProvider(resolutionPath, contextSchemaPath, owner, referencedAssemblies) :> ISqlProvider
@@ -31,19 +33,21 @@ type public SqlDataContext (typeName, connectionString:string, providerType, res
     let myLock2 = new Object();
 
     let provider =
-        providerCache.GetOrAdd(typeName,
-            fun typeName ->
-                lazy 
-                    let prov : ISqlProvider = ProviderBuilder.createProvider providerType resolutionPath referencedAssemblies runtimeAssembly owner tableNames contextSchemaPath odbcquote sqliteLibrary
-                    if not (prov.GetSchemaCache().IsOffline) then
-                        use con = prov.CreateConnection(connectionString)
-                        con.Open()
-                        // create type mappings and also trigger the table info read so the provider has
-                        // the minimum base set of data available
-                        prov.CreateTypeMappings(con)
-                        prov.GetTables(con,caseSensitivity) |> ignore
-                        if (providerType <> DatabaseProviderTypes.MSACCESS && providerType.GetType() <> typeof<Providers.MSAccessProvider>) then con.Close()
-                    prov).Value
+        let addCache() =
+            lazy
+                let prov : ISqlProvider = ProviderBuilder.createProvider providerType resolutionPath referencedAssemblies runtimeAssembly owner tableNames contextSchemaPath odbcquote sqliteLibrary
+                if not (prov.GetSchemaCache().IsOffline) then
+                    use con = prov.CreateConnection(connectionString)
+                    con.Open()
+                    // create type mappings and also trigger the table info read so the provider has
+                    // the minimum base set of data available
+                    prov.CreateTypeMappings(con)
+                    prov.GetTables(con,caseSensitivity) |> ignore
+                    if (providerType <> DatabaseProviderTypes.MSACCESS && providerType.GetType() <> typeof<Providers.MSAccessProvider>) then con.Close()
+                prov
+        try providerCache.GetOrAdd(typeName, fun _ -> addCache()).Value
+        with | _ -> providerCache.AddOrUpdate(typeName, addCache(), fun _ _ -> addCache()).Value
+
 
     let initCallSproc (dc:ISqlDataContext) (def:RunTimeSprocDefinition) (values:obj array) (con:IDbConnection) (com:IDbCommand) =